requirements.txt
typer[all]
rich
requests
dnspython
